move sentinel out of IPython.utils.signature
signature is a backport of a stdlib file,
and not an appropriate place for new code.

Further, to avoid adding a new trivial module to genutils,
a copy is added each place we've used a sentinel
